requirements.txt
pytesseract==0.3.13
opencv-python==4.11.0.86
scikit-image==0.25.2
numpy==2.2.4
pandas==2.2.3
python-dotenv==1.1.0
pywavelets==1.8.0
Pillow==11.2.0
scipy==1.15.2